requirements.txt
flask
langchain-groq
langchain
langgraph
python-dotenv
duckduckgo-search